add anonymous mode in user history && input empty string will do nothing

+++ frontend.py
@@ -3,7 +3,7 @@ import operator
 import json
 import httplib2
 import requests
-from backend.data_structures import UserHistoryIndex, History
+from backend.data_structures import UserHistoryIndex, UserSession, History
 from bottle import Bottle, route, run, template, get, post, request, static_file, redirect, app
 from oauth2client.client import OAuth2WebServerFlow, flow_from_clientsecrets
 from googleapiclient.errors import HttpError
@@ -13,19 +13,12 @@ from beaker.middleware import SessionMiddleware
 
 # declare golbal variables
 login = False
-'''
-token = {}
-session = {}
-user_document = {}
-user_email = ''
-user_name = ''
-'''
 user_history_index = UserHistoryIndex()
-
+user_session = UserSession()
 current_page = 'query_page'
-keywords  = ''
+keywords = ''
 words_count = []
-
+current_user_email = 'anonymous_mode'
 SCOPE = 'https://www.googleapis.com/auth/plus.me https://www.googleapis.com/auth/userinfo.email'
 REDIRECT_URI = 'http://localhost:8081/redirect'
 sessions_opts = {
@@ -36,26 +29,30 @@ sessions_opts = {
 }
 
 # ask for keywords from user
+
+
 @get('/')  # or @route('/')
 def home():
     print "------route---home------------------------------"
 
     if login:
-        session = request.environ.get('beaker.session')
-        return template('./templates/query_page.tpl', login = login, 
-            user_email = session["user_email"], history = user_history_index.get_history(session["user_email"]).get_popular())
+        #session = request.environ.get('beaker.session')
+        return template('./templates/query_page.tpl', login=login,
+                        user_email=session["user_email"], history=user_history_index.get_history(session["user_email"]).get_popular())
     else:
-        return template('./templates/query_page.tpl', login = login)
+        return template('./templates/query_page.tpl', login=login,
+                        history=user_history_index.get_history('anonymous_mode').get_popular())
 
 # show search results, word count, and search history
+
+
 @post('/')  # or @route('/', method='POST')
 def show_results():
-    
+
     # keyword from http get
     global keywords
     global words_count
     keywords = request.forms.get('keywords')
-
     # split keyword string into words and count them
     # store words in a dict
     words_list = keywords.split()
@@ -64,23 +61,34 @@ def show_results():
     # add keyword to history
     # joining words instead of the original string to avoid multiple whitespaces
     if login:
-        session = request.environ.get('beaker.session')
-        user_history_index.get_history(session["user_email"]).add_new_keywords(words_list)
+        new_session = request.environ.get('beaker.session')
+        if new_session:
+            email = new_session['user_email']
+        else:
+            email = session['user_email']
+        user_history_index.get_history(email).add_new_keywords(words_list)
         print "words_count: "+repr(words_count)
-        return template('./templates/result_page.tpl', keywords = keywords, words_count = words_count, login = login, 
-            user_email = session["user_email"], history = user_history_index.get_history(session["user_email"]).get_popular())
+        return template('./templates/result_page.tpl', keywords=keywords, words_count=words_count, login=login,
+                        user_email=email, history=user_history_index.get_history(email).get_popular())
     else:
-        return template('./templates/result_page.tpl', keywords = keywords, words_count = words_count, login = login)
+        user_history_index.get_history(
+            'anonymous_mode').add_new_keywords(words_list)
+        return template('./templates/result_page.tpl', keywords=keywords, words_count=words_count, login=login,
+                        history=user_history_index.get_history('anonymous_mode').get_popular())
 
-#if user login in the query_page, set the current page to query_page
+# if user login in the query_page, set the current page to query_page
+
+
 @route('/login', method='GET')
 def query_page():
     global current_page
     current_page = 'query_page'
     google_login()
-     
-#if user login in the result_page, set the current page to result_page
-#and then redirect to Google login
+
+# if user login in the result_page, set the current page to result_page
+# and then redirect to Google login
+
+
 @route('/login/result', method='GET')
 def result_page():
     global current_page
@@ -88,16 +96,20 @@ def result_page():
     google_login()
 
 # redirect to Google login prompt for user authentication
+
+
 def google_login():
-    
+
     print "------route---login------------------------------"
-    flow = flow_from_clientsecrets('client_secrets.json', scope = SCOPE, redirect_uri = REDIRECT_URI)
+    flow = flow_from_clientsecrets(
+        'client_secrets.json', scope=SCOPE, redirect_uri=REDIRECT_URI)
     uri = flow.step1_get_authorize_url()
     print "flow: "
     print flow
     print "uri: "
     print uri
     return redirect(str(uri))
+
 
 def credentials_to_dict(credentials):
     return {'access_token': credentials.access_token,
@@ -109,29 +121,34 @@ def credentials_to_dict(credentials):
             'scopes': credentials.scopes}
 
 # redirect to Google logout prompt, and then redirect to the query_page
+
+
 @route('/logout', method='GET')
 def google_logout():
     print "------route---logout------------------------------"
-    #TODO: empty gobal variables
+    # TODO: empty gobal variables
     global login
     login = False
 
-    session = request.environ.get('beaker.session')
-    requests.post('https://accounts.google.com/o/oauth2/revoke',
-        params={'token': session["token"]},
-        headers = {'content-type': 'application/x-www-form-urlencoded'})
+    session = user_session.get_session(current_user_email)
+    if session != {}:
+        r = requests.post('https://accounts.google.com/o/oauth2/revoke',
+                          params={'token': session["token"]},
+                          headers={'content-type': 'application/x-www-form-urlencoded'})
+        print "\n requests.post " + repr(r)+"\n"
 
-    print "\n session: " + repr(session)
-    session.delete()
-    print "\n session: " + repr(session)
- 
+        session.delete()
+
     return redirect('/')
+
 
 '''
 If user authorizes your application server to access the Google services, an one-time code will be attached
 to the query string when the browser is redirected to the redirect_uri specified in step 2. 
 The one-time code can be retrieved as GET parameter:
 '''
+
+
 @route('/redirect')
 def redirect_page():
     print "------route---redirect------------------------------"
@@ -144,7 +161,7 @@ def redirect_page():
         print load_dict
     print "2\n"
     flow = OAuth2WebServerFlow(client_id=load_dict['web']['client_id'], client_secret=load_dict['web']['client_secret'],
-                                scope=SCOPE, redirect_uri=REDIRECT_URI)
+                               scope=SCOPE, redirect_uri=REDIRECT_URI)
     credentials = flow.step2_exchange(code)
 
     print credentials_to_dict(credentials)
@@ -164,36 +181,32 @@ def redirect_page():
     # Get user info
     users_service = build('oauth2', 'v2', http=http)
     user_document = users_service.userinfo().get().execute()
-    #print "\n user_document: " + repr(user_document)
+    # print "\n user_document: " + repr(user_document)
     user_email = user_document['email']
 
-    global user_name
-    if not user_document['name']:
-        user_name=user_document['name']
-
-    #if it is a new user, add the user to class
-    global user
-    global user_history
-    user.add_if_new_user(user_email)
-    user_history = user.get_history(user_email)
-
-
-    # maintain a session for the user
-    global session
-    #print "\n environ: " + repr(request.environ)
+    # print "\n environ: " + repr(request.environ)
     # Fetch user history from database or create a new one
     # maintain a session for the user
     session = request.environ.get('beaker.session')
     session["user_email"] = user_email
     session['credentials'] = credentials_to_dict(credentials)
     session["token"] = token
-
+    if "name" in user_document.keys():
+        session['user_name'] = user_document['name']
     session.save()
-    
+
     print "\n session: " + repr(session)
+
+    global current_user_email
+    current_user_email = user_email
+    global user_session
+    user_session.add_new_user_session(user_email, session)
+
     return redirect('/user')
 
-#after user login, they will stay on the same page (query_page || result_page)
+# after user login, they will stay on the same page (query_page || result_page)
+
+
 @route('/user')
 def user_login():
     print "------route---user------------------------------"
@@ -203,12 +216,12 @@ def user_login():
     session = request.environ.get('beaker.session')
 
     if current_page == 'query_page':
-        return template('./templates/query_page.tpl', login = login, 
-            user_email = session["user_email"], history = user_history_index.get_history(session["user_email"]).get_popular())
+        return template('./templates/query_page.tpl', login=login,
+                        user_email=current_user_email, history=user_history_index.get_history(current_user_email.get_popular())
     else:
-        return template('./templates/result_page.tpl', keywords = keywords, words_count = words_count, login = login, 
-            user_email = session["user_email"], history = user_history_index.get_history(session["user_email"]).get_popular())
-       
+        return template('./templates/result_page.tpl', keywords=keywords, words_count=words_count, login=login,
+                        user_email=current_user_email, history=user_history_index.get_history(current_user_email.get_popular())
+
 
 # routes of assets (css, js, images)
 @route('/assets/<filename:path>')
@@ -216,9 +229,12 @@ def send_assets(filename):
     return static_file(filename, root='./assets')
 
 # route of templates
+
+
 @route('/templates/<filename:path>')
 def send_templates(filename):
     return static_file(filename, root='./templates')
+
 
 if __name__ == "__main__":
     app = SessionMiddleware(app(), sessions_opts)
